+++ src/main/java/com/olliego/community/model/Question.java
@@ -1,7 +1,7 @@
 package com.olliego.community.model;
 
 public class Question {
-    private Integer id;
+    private Long id;
 
     private String title;
 
@@ -9,23 +9,23 @@ public class Question {
 
     private Long gmtModified;
 
-    private Integer creator;
+    private Long creator;
 
-    private Integer commentCount;
+    private Long commentCount;
 
-    private Integer viewCount;
+    private Long viewCount;
 
-    private Integer likeCount;
+    private Long likeCount;
 
     private String tag;
 
     private String description;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -53,35 +53,35 @@ public class Question {
         this.gmtModified = gmtModified;
     }
 
-    public Integer getCreator() {
+    public Long getCreator() {
         return creator;
     }
 
-    public void setCreator(Integer creator) {
+    public void setCreator(Long creator) {
         this.creator = creator;
     }
 
-    public Integer getCommentCount() {
+    public Long getCommentCount() {
         return commentCount;
     }
 
-    public void setCommentCount(Integer commentCount) {
+    public void setCommentCount(Long commentCount) {
         this.commentCount = commentCount;
     }
 
-    public Integer getViewCount() {
+    public Long getViewCount() {
         return viewCount;
     }
 
-    public void setViewCount(Integer viewCount) {
+    public void setViewCount(Long viewCount) {
         this.viewCount = viewCount;
     }
 
-    public Integer getLikeCount() {
+    public Long getLikeCount() {
         return likeCount;
     }
 
-    public void setLikeCount(Integer likeCount) {
+    public void setLikeCount(Long likeCount) {
         this.likeCount = likeCount;
     }
