+++ src/main/java/com/olliego/community/service/QuestionService.java
@@ -74,7 +74,7 @@ public class QuestionService {
         return paginationDTO;
     }
 
-    public PaginationDTO list(Integer userId, Integer page, Integer size) {
+    public PaginationDTO list(Long userId, Integer page, Integer size) {
         PaginationDTO paginationDTO = new PaginationDTO();
         Integer totalPage;
 
@@ -122,7 +122,7 @@ public class QuestionService {
         return paginationDTO;
     }
 
-    public QuestionDTO getById(Integer id) {
+    public QuestionDTO getById(Long id) {
         Question question = questionMapper.selectByPrimaryKey(id);
         if (question == null) {
             throw new CustomizeException("你要找的问题不在了");
@@ -138,9 +138,9 @@ public class QuestionService {
         if(question.getId() == null){
             question.setGmtCreate(System.currentTimeMillis());
             question.setGmtModified(question.getGmtCreate());
-            question.setLikeCount(0);
-            question.setViewCount(0);
-            question.setCommentCount(0);
+            question.setLikeCount(0L);
+            question.setViewCount(0L);
+            question.setCommentCount(0L);
             questionMapper.insert(question);
         }else {
             Question updateQuestion = new Question();
@@ -158,10 +158,10 @@ public class QuestionService {
         }
     }
 
-    public void incView(Integer id) {
+    public void incView(Long id) {
         Question question = new Question();
         question.setId(id);
-        question.setViewCount(1);
+        question.setViewCount(1L);
         questionExtMapper.incView(question);
     }
 }
